requirements.txt
# Détection Automatique de l'Intensité de Douleur
# Python 3.8 - 3.11 requis

# Vision & Deep Learning
opencv-python==4.10.0.84
mediapipe==0.10.14

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Interface (optionnel)
streamlit==1.28.0

# Utilities
Pillow==10.0.0